requirements.txt
streamlit
requests
python-dotenv
jinja2
deepl
msal_streamlit_authentication
